Majorly cleanup reactivity module
Moved a lot of `state` files into a single core.rs file, and fixed up
techical debt in naming and docs.
